FIR IDE: Enable RemoveUselessIsCheckFix and RemoveUselessIsCheckFixForWhen.
