+++ idea/testData/quickfix/expressions/removeUselessIsCheckInWhen.kt.after
@@ -8,5 +8,3 @@ fun foo(bar: Base):Int {
         else -> 42
     }
 }
-
-/* IGNORE_FIR */
