+++ idea/testData/quickfix/expressions/removeUselessIsCheckInWhenNegate.kt
@@ -9,5 +9,3 @@ fun foo(bar: Base):Int {
         else -> 1
     }
 }
-
-/* IGNORE_FIR */
