+++ idea/testData/quickfix/expressions/removeUselessIsCheckInWhenNegate.kt.after
@@ -8,5 +8,3 @@ fun foo(bar: Base):Int {
         else -> 1
     }
 }
-
-/* IGNORE_FIR */
