+++ idea/testData/quickfix/expressions/removeUselessIsCheckNegate.kt
@@ -4,5 +4,3 @@ fun foo(a: String) {
 
     }
 }
-
-/* IGNORE_FIR */
